+++ CHANGELOG.md
@@ -10,6 +10,7 @@
   - Fetch the details for the given item id from the given collection id
   - Add tags to a job
   - Create a collection item
+  - Create a link between 2 collection items
 - Fix handling of 404 HTTP status.
 
 ## [0.3.1] - 2022-11-30