+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # CHANGELOG
 
-## [0.3.2]
+## [0.3.2] - 2023-04-23
 
 - Add new requests:
   - Upload activestorage attachment